+++ checker/jtreg/unboundedWildcards/issue1427/T.java
@@ -4,7 +4,7 @@
  * https://github.com/typetools/checker-framework/issues/1427
  *
  * @compile B.java
- * @compile -XDrawDiagnostics -processor org.checkerframework.checker.tainting.TaintingChecker T.java
+ * @compile -processor org.checkerframework.checker.tainting.TaintingChecker T.java
  */
 public class T {
   {