+++ dataflow/src/main/java/org/checkerframework/dataflow/cfg/visualize/AbstractCFGVisualizer.java
@@ -244,8 +244,8 @@ public abstract class AbstractCFGVisualizer<
 
     List<Node> contents = addBlockContent(bb);
     StringJoiner sjBlockContents = new StringJoiner(separator);
-    for (Node t : contents) {
-      sjBlockContents.add(visualizeBlockNode(t, analysis));
+    for (Node n : contents) {
+      sjBlockContents.add(visualizeBlockNode(n, analysis));
     }
     return sjBlockContents.toString();
   }
@@ -261,22 +261,34 @@ public abstract class AbstractCFGVisualizer<
   }
 
   /**
-   * Format the given object as a String suitable for the output format, i.e. with format-specific
-   * characters escaped.
+   * Make the string suitable for the output format, i.e. with format-specific characters escaped.
+   *
+   * @param str a string
+   * @return the escaped string
+   */
+  protected String escapeString(String str) {
+    return str;
+  }
+
+  /**
+   * Make the object's string representation suitable for the output format, i.e. with
+   * format-specific characters escaped.
    *
    * @param obj an object
-   * @return the formatted String from the given object
+   * @return the object's string representation, escaped
    */
-  protected abstract String format(Object obj);
+  protected final String escapeString(Object obj) {
+    return escapeString(obj.toString());
+  }
 
   @Override
-  public String visualizeBlockNode(Node t, @Nullable Analysis<V, S, T> analysis) {
+  public String visualizeBlockNode(Node n, @Nullable Analysis<V, S, T> analysis) {
     StringBuilder sbBlockNode = new StringBuilder();
-    sbBlockNode.append(format(t)).append("   [ ").append(getNodeSimpleName(t)).append(" ]");
+    sbBlockNode.append(escapeString(n)).append("   [ ").append(getNodeSimpleName(n)).append(" ]");
     if (analysis != null) {
-      V value = analysis.getValue(t);
+      V value = analysis.getValue(n);
       if (value != null) {
-        sbBlockNode.append("    > ").append(format(value));
+        sbBlockNode.append("    > ").append(escapeString(value));
       }
     }
     return sbBlockNode.toString();
@@ -474,11 +486,11 @@ public abstract class AbstractCFGVisualizer<
   /**
    * Get the simple name of a node.
    *
-   * @param t a node
+   * @param n a node
    * @return the node's simple name, without "Node"
    */
-  protected String getNodeSimpleName(Node t) {
-    String name = t.getClass().getSimpleName();
+  protected String getNodeSimpleName(Node n) {
+    String name = n.getClass().getSimpleName();
     return name.replace("Node", "");
   }
 }