+++ javacutil/src/main/java/org/checkerframework/javacutil/TreeUtils.java
@@ -2576,14 +2576,14 @@ public final class TreeUtils {
    * @param tree a method/constructor invocation
    * @return true if the given method/constructor invocation is a varargs invocation
    */
-  public static boolean isVarArgs(Tree tree) {
+  public static boolean isVarargsCall(Tree tree) {
     switch (tree.getKind()) {
       case METHOD_INVOCATION:
-        return isVarArgs((MethodInvocationTree) tree);
+        return isVarargsCall((MethodInvocationTree) tree);
       case NEW_CLASS:
-        return isVarArgs((NewClassTree) tree);
+        return isVarargsCall((NewClassTree) tree);
       default:
-        throw new BugInCF("TreeUtils.isVarArgs: unexpected kind of tree: " + tree);
+        return false;
     }
   }
 
@@ -2592,9 +2592,21 @@ public final class TreeUtils {
    *
    * @param invok the method invocation
    * @return true if the given method invocation is a varargs invocation
+   * @deprecated use {@link #isVarargsCall(MethodInvocationTree)}
    */
+  @Deprecated // 2024-06-04
   public static boolean isVarArgs(MethodInvocationTree invok) {
-    return isVarArgs(elementFromUse(invok), invok.getArguments());
+    return ((JCMethodInvocation) invok).varargsElement != null;
+  }
+
+  /**
+   * Returns true if the given method invocation is a varargs invocation.
+   *
+   * @param invok the method invocation
+   * @return true if the given method invocation is a varargs invocation
+   */
+  public static boolean isVarargsCall(MethodInvocationTree invok) {
+    return ((JCMethodInvocation) invok).varargsElement != null;
   }
 
   /**
@@ -2620,38 +2632,21 @@ public final class TreeUtils {
    *
    * @param newClassTree the constructor invocation
    * @return true if the given method invocation is a varargs invocation
+   * @deprecated use {@link #isVarargsCall(NewClassTree)}
    */
+  @Deprecated // 2024-06-04
   public static boolean isVarArgs(NewClassTree newClassTree) {
-    return isVarArgs(elementFromUse(newClassTree), newClassTree.getArguments());
+    return isVarargsCall(newClassTree);
   }
 
   /**
-   * Returns true if a method/constructor invocation is a varargs invocation.
+   * Returns true if the given constructor invocation is a varargs invocation.
    *
-   * @param method the method or constructor
-   * @param args the arguments passed at the invocation
-   * @return true if the given method/constructor invocation is a varargs invocation
+   * @param newClassTree the constructor invocation
+   * @return true if the given method invocation is a varargs invocation
    */
-  private static boolean isVarArgs(ExecutableElement method, List<? extends ExpressionTree> args) {
-    if (!method.isVarArgs()) {
-      return false;
-    }
-
-    List<? extends VariableElement> parameters = method.getParameters();
-    if (parameters.size() != args.size()) {
-      return true;
-    }
-
-    TypeMirror lastArgType = typeOf(args.get(args.size() - 1));
-    if (lastArgType.getKind() == TypeKind.NULL) {
-      return false;
-    }
-    if (lastArgType.getKind() != TypeKind.ARRAY) {
-      return true;
-    }
-
-    TypeMirror varargsParamType = parameters.get(parameters.size() - 1).asType();
-    return TypesUtils.getArrayDepth(varargsParamType) != TypesUtils.getArrayDepth(lastArgType);
+  public static boolean isVarargsCall(NewClassTree newClassTree) {
+    return ((JCNewClass) newClassTree).varargsElement != null;
   }
 
   /**
@@ -2842,15 +2837,11 @@ public final class TreeUtils {
    * @param methodInvocation a method or constructor invocation
    * @return whether applicability by variable arity invocation is necessary to determine the method
    *     signature
+   * @deprecated use {@link #isVarargsCall(Tree)}
    */
+  @Deprecated // 2024-06-04
   public static boolean isVarArgMethodCall(ExpressionTree methodInvocation) {
-    if (methodInvocation.getKind() == Tree.Kind.METHOD_INVOCATION) {
-      return ((JCMethodInvocation) methodInvocation).varargsElement != null;
-    } else if (methodInvocation.getKind() == Tree.Kind.NEW_CLASS) {
-      return ((JCNewClass) methodInvocation).varargsElement != null;
-    } else {
-      return false;
-    }
+    return isVarargsCall(methodInvocation);
   }
 
   /**